Changed doubles

+++ Homework/Assignment 2/Savitch_9thEd_Ch2_Prob12_Prog_Proj/main.cpp
@@ -17,25 +17,25 @@
 
 using namespace std;
 
-const double PI = 3.14;
-const double GAL_IN_FT = 7.48;
+const float PI = 3.14f;
+const float GAL_IN_FT = 7.48f;
 
 int main(int argc, char** argv) {
     
-    double rad;
-    double depth;
-    double cubeFt;
+    float rad;
+    float depth;
+    float cubeFt;
     cout << fixed << setprecision(2);
     
     cout << "Enter the radius of the well casing in inches.\n";
     cin >> rad;
-    double radFt = rad / 12;
+    float radFt = rad / 12;
     cout << "Enter the depth of the well in feet.\n";
     cin >> depth;
     cout << endl;
     cubeFt = PI * pow(radFt,2) * depth;
     cout << "There are " << cubeFt << " cubic ft. in this well\n";
-    double galInWl = cubeFt * GAL_IN_FT;
+    float galInWl = cubeFt * GAL_IN_FT;
     cout << fixed << setprecision(0);
     
     if (galInWl > 250){
